+++ src/components/searchField/searchField.tsx
@@ -9,7 +9,7 @@ import {
   TextField,
   Typography,
 } from '@mui/material';
-import React, { useState } from 'react';
+import { useState } from 'react';
 import ArrowUpDownIcon from '../../assets/arrowUpDownIcon';
 import ClockIcon from '../../assets/clockIcon';
 import Close from '../../assets/close';
@@ -18,7 +18,6 @@ import EscapeIcon from '../../assets/escapeIcon';
 import HoverEnter from '../../assets/hoverEnter';
 import Search from '../../assets/search';
 import './index.css';
-import { styles } from './styles';
 import {
   Category1Props,
   ControlsProps,
@@ -26,239 +25,8 @@ import {
   SearchFieldProps,
   SuggestionProp,
 } from './props';
-
-const SuggestionRecentSearch = (props: SuggestionProp) => {
-  const { searchValue, option, handleRecentSearchData = () => {} } = props;
-
-  const filteredRows = option.filter((row: any) => {
-    return row.label.toLowerCase().includes(searchValue?.toLowerCase());
-  });
-  return (
-    <>
-      {filteredRows?.map((item: any) => (
-        <Stack
-          direction={'row'}
-          columnGap={'8px'}
-          rowGap={'24px'}
-          p={'0 14px'}
-          marginBottom={'16px'}
-          onClick={() => handleRecentSearchData(item)}
-        >
-          <>
-            <Box>
-              <ClockIcon />
-            </Box>
-            <Box>
-              <Typography color={'#262626'} fontSize={'14px'} fontWeight={600}>
-                {item?.label}
-              </Typography>
-            </Box>
-          </>
-        </Stack>
-      ))}{' '}
-    </>
-  );
-};
-
-const FoodCategory = (props: Category1Props) => {
-  const { searchValue, option, handleCategoryData = () => {} } = props;
-
-  const filteredRows = option.filter((row: any) => {
-    return row.label.toLowerCase().includes(searchValue?.toLowerCase());
-  });
-
-  return (
-    <>
-      <Stack direction={'row'} overflow={'scroll'}>
-        {option?.map((val: any) => (
-          <Stack
-            {...props}
-            sx={{
-              p: '14px',
-              cursor: 'pointer',
-            }}
-            direction={'column'}
-            onClick={() => handleCategoryData(val)}
-          >
-            <img
-              src={val?.url}
-              alt=" "
-              style={{
-                width: '74px',
-                height: '74px',
-                borderRadius: '50%',
-              }}
-            />
-            <Typography
-              color={'#3B3B3B'}
-              fontSize={'12px'}
-              mt={'6px'}
-              width={'74px'}
-              className="title1"
-              textAlign={'center'}
-            >
-              {val?.label}
-            </Typography>
-          </Stack>
-        ))}
-      </Stack>
-    </>
-  );
-};
-
-const EmployeeCategory = (props: Category1Props) => {
-  const { searchValue, option, handleCategoryData = () => {} } = props;
-
-  const filteredRows = option.filter((row: any) => {
-    return row?.label?.toLowerCase()?.includes(searchValue?.toLowerCase());
-  });
-  console.log(option);
-
-  return (
-    <>
-      {option?.map((val: any) => (
-        <Stack
-          {...props}
-          sx={{
-            p: '14px',
-            cursor: 'pointer',
-          }}
-          direction={'row'}
-          alignItems={'center'}
-          onClick={() => handleCategoryData(val)}
-        >
-          <Box>
-            <img
-              src={val?.url}
-              alt=" "
-              style={{
-                width: '40px',
-                height: '40px',
-                borderRadius: '4px',
-              }}
-            />
-          </Box>
-          <Box ml={'8px'}>
-            <Typography
-              color={'#3B3B3B'}
-              fontSize={'12px'}
-              fontWeight={600}
-              className="title1"
-            >
-              {val?.label}
-            </Typography>
-            <Typography color={'#666666'} fontSize={'10px'} mt={'6px'}>
-              {val?.label}
-            </Typography>
-          </Box>
-        </Stack>
-      ))}
-    </>
-  );
-};
-
-const ControlSearch = (props: ControlsProps) => {
-  const {
-    navigateButtons,
-    controlsRootStyles,
-    controlsBgColor,
-    controlsTextSize,
-    controlsTextColor,
-  } = props;
-  return (
-    <Stack
-      style={{
-        // ...controlsRootStyles,
-        cursor: 'pointer',
-        minHeight: '34px',
-        padding: '9px',
-        borderBottomRightRadius: '8px',
-        borderBottomLeftRadius: '8px',
-      }}
-      direction={'row'}
-      alignItems={'center'}
-      columnGap={'20px'}
-      bgcolor={controlsBgColor}
-    >
-      {navigateButtons?.map((val: any) => (
-        <Stack
-          direction={'row'}
-          alignItems={'center'}
-          onClick={() => val?.onClick()}
-        >
-          {val?.icon}
-          <Typography
-            ml={'6px'}
-            color={controlsTextColor}
-            fontSize={controlsTextSize}
-          >
-            {val?.navigateLabel}
-          </Typography>
-        </Stack>
-      ))}
-    </Stack>
-  );
-};
-
-const RecentSearch = (props: RecentSearchProps) => {
-  const {
-    recentVal,
-    recentLabel,
-    recentLabelTextColor,
-    recentLabelFontSize,
-    recentLabelStyle,
-    recentSearchItemStyles,
-    recentSearchItemTextColor,
-    recentSearchItemBgcolor,
-    recentSearchItemSize,
-  } = props;
-
-  return (
-    <Stack
-      style={{
-        cursor: 'pointer',
-        padding: '9px',
-      }}
-    >
-      <Typography
-        color={recentLabelTextColor}
-        fontSize={recentLabelFontSize}
-        sx={{ ...recentLabelStyle }}
-      >
-        {recentLabel}
-      </Typography>
-      <Stack direction={'row'} mt={'8px'} mb={1} overflow={'scroll'}>
-        {recentVal?.map((val: any) => (
-          <>
-            <Typography
-              fontWeight={600}
-              padding={'4px 6px'}
-              mr={'12px'}
-              color={recentSearchItemTextColor}
-              fontSize={recentSearchItemSize}
-              bgcolor={recentSearchItemBgcolor}
-              borderRadius={'8px'}
-              minWidth={'fit-content'}
-              sx={{ ...recentSearchItemStyles }}
-            >
-              {val?.label}
-            </Typography>
-          </>
-        ))}
-      </Stack>
-      <Divider sx={{ mt: '0px' }} />
-    </Stack>
-  );
-};
-
-interface handleSearchDataProp {
-  label: string;
-}
-interface dataType {
-  label: string;
-  title: string;
-}
-[];
+import { styles } from './styles';
+import { ControlSearch, EmployeeCategory, FoodCategory, SuggestionRecentSearch } from './helperComponents';
 
 const SearchField = (props: SearchFieldProps) => {
   const {
@@ -315,9 +83,20 @@ const SearchField = (props: SearchFieldProps) => {
     controlsTextColor,
     controlsTextSize,
 
+    componentColumnDirection,
+    recentSearchLabelColor,
+    recentSearchLabelSize,
+    recentSearchLabel,
     categoryLabel1,
     categoryLabel2,
+    categoryLabel1Color,
+    categoryLabel1Size,
+    categoryLabel2Color,
+    categoryLabel2Size,
+
     textSearchData,
+    navigateData,
+    cardData,
   } = props;
 
   const [searched, setSearched] = useState(null);
@@ -327,12 +106,6 @@ const SearchField = (props: SearchFieldProps) => {
 
   const requestOnInputSearch = (val: string) => {
     setSearch(val);
-  };
-
-  const handleKeyDown = (event: KeyboardEvent) => {
-    if (event.key === '') {
-      setOpen(true);
-    }
   };
 
   const handleSearchData = (val: any) => {
@@ -359,14 +132,14 @@ const SearchField = (props: SearchFieldProps) => {
         sx={{
           minWidth: '100%',
         }}
-        // onClick={() => handleOpen()}
+        onClick={() => handleOpen()}
       >
         <Autocomplete
           open={isOpen}
           clearOnEscape={true}
           fullWidth
           value={searched}
-          options={textSearchData}
+          options={cardData}
           onChange={(e, value) => handleOnchange(e, value)}
           getOptionLabel={(option) => option?.label}
           sx={{
@@ -609,6 +382,7 @@ const SearchField = (props: SearchFieldProps) => {
                   />
                 </Box>
               )}
+
               {children}
 
               {isShortcutKeyBased && (
@@ -628,13 +402,13 @@ const SearchField = (props: SearchFieldProps) => {
                   <Divider />
                   <Stack direction={'column'}>
                     <Typography
-                      color={'#929292'}
-                      fontSize={'12px'}
+                      color={recentSearchLabelColor}
+                      fontSize={recentSearchLabelSize}
                       mt={'24px'}
                       mb={'12px'}
                       pl={'16px'}
                     >
-                      RECENT SEARCH
+                      {recentSearchLabel}
                     </Typography>
                   </Stack>
                   <SuggestionRecentSearch
@@ -643,11 +417,11 @@ const SearchField = (props: SearchFieldProps) => {
                   />
 
                   <Divider />
-                  <Stack direction={'column-reverse'}>
+                  <Stack direction={componentColumnDirection}>
                     <Stack>
                       <Typography
-                        color={'#929292'}
-                        fontSize={'12px'}
+                        color={categoryLabel1Color}
+                        fontSize={categoryLabel1Size}
                         mt={'24px'}
                         pl={'16px'}
                       >
@@ -661,8 +435,8 @@ const SearchField = (props: SearchFieldProps) => {
                     <Divider />
                     <Stack>
                       <Typography
-                        color={'#929292'}
-                        fontSize={'12px'}
+                        color={categoryLabel2Color}
+                        fontSize={categoryLabel2Size}
                         mt={'24px'}
                         pl={'16px'}
                       >
@@ -670,7 +444,7 @@ const SearchField = (props: SearchFieldProps) => {
                       </Typography>
                       <EmployeeCategory
                         option={cardData}
-                        handleEmployeeSearchData={handleSearchData}
+                        handleCategoryData={handleSearchData}
                       />
                     </Stack>
                   </Stack>
@@ -817,7 +591,17 @@ SearchField.defaultProps = {
   controlsBgColor: '#E9E9E9',
   controlsTextSize: 12,
 
+  componentColumnDirection: '',
+
+  recentSearchLabelColor: '#393939dd',
+  recentSearchLabelSize: 12,
+  recentSearchLabel: 'Recent Search',
+
+  categoryLabel1Color: '#393939dd',
+  categoryLabel1Size: 12,
   categoryLabel1: ' CATEGORY SUGGEST 1',
+  categoryLabel2Color: '#393939dd',
+  categoryLabel2Size: 12,
   categoryLabel2: 'CATEGORY SUGGEST 2',
 
   textSearchData: [
